app/backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data Validation
pydantic>=2.5.3

# Scheduler
apscheduler>=3.10.4

# RSS Parsing
feedparser>=6.0.10

# HTTP Client
httpx>=0.26.0

# Web Scraping
beautifulsoup4>=4.12.3
lxml>=5.1.0

# Firecrawl - AI-powered web scraping
firecrawl-py>=1.0.0

# Database (SQLite built-in, aiosqlite for async if needed)
aiosqlite>=0.19.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
hypothesis>=6.92.2

# Utilities
python-dateutil>=2.8.2
